Add KVM_SET_TSC_KHZ KVM_GET_TSC_KHZ for x86
The KVM_GET_TSC_KHZ and KVM_SET_TSC_KHZ allows users to access and
control the frequency of the Time Stamp Counter. This is useful in
cases such as snapshotting, where resuming from a snapshot on a
different CPU than on the original machine will lead to inconsitencies
in applications which make use of the Time Stamp Counter.

This commit:
- implements vCPU wrappers: set_tsc_khz(), get_tsc_khz()
- adds unit tests
- tests for expected failures if KVM_CAP_TSC_CONTROL or
KVM_CAP_GET_TSC_KHZ are missing

Signed-off-by: Andrei Sandu <sandreim@amazon.com>
Signed-off-by: George Pisaltu <gpl@amazon.com>