ioctls: Don't panic on unsupported exit reason
In case the underlying KVM version is more recent than what is known
and supported by the kvm-ioctls crate, and in case KVM reports an
unknown reason for a VM exit, we don't want the vCPU run() to panic.

A more elegant way of handling such situation is by propagating the exit
reason up to the consumer's crate as it might know what to do with it.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>