Add support for userspace MSR handling
Add the appropriate types to allow for a VMM to intercept RDMSR and
WRMSR instructions if access to an MSR is denied. By default, KVM
injects #GP on denied accesses.

To do this we need to add a new Cap variant which corresponds to
KVM_CAP_X86_USER_SPACE_MSR, and two new VcpuExit variants that
correspond to KVM_EXIT_X86_RDMSR and KVM_EXIT_X86_WRMSR respectively.
We also need helper types to pass the relevant information about the
exit to the VMM.

Signed-off-by: Carlos López <carlos.lopez@suse.com>