Mark DeviceFd::get_device_attr as unsafe
KVM_GET_DEVICE_ATTR causes the kernel to write to an arbitrary
address.  This is unsafe, as it can allow writing to an address Rust
believes to be immutable.

I discovered this because an optimisation change[1] in Rust 1.80.0
caused a Cloud Hypervisor test to start failing when built in release
mode, because it was setting the addr passed to get_device_attr() to
the address of an immutable variable.

[1]: https://github.com/rust-lang/rust/commit/d2d24e395a1e4fcee62ca17bf4cbddb1f903af97

Fixes: 744e8bd ("Add support for `KVM_GET_DEVICE_ATTR` ioctl")
Signed-off-by: Alyssa Ross <hi@alyssa.is>